Style pages to look like GOV.UK
Use Slimmer (and hence Static) to add the GOV.UK header and footer to our
pages.

Also use the title govuk_component to render the h1 in the GOV.UK style.

Use the core_layout to avoid loading the styles for many formats that this app
doesn't provide, as is the recommended style.